refactor: useCoupon 커스텀 훅 구현하여 비즈니스 로직 분리
